Split out custom SQL menu items
