+++ src/vs/platform/actions/browser/menuEntryActionViewItem.ts
@@ -4,11 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import 'vs/css!./menuEntryActionViewItem';
-import { asCSSUrl, createCSSRule, ModifierKeyEmitter } from 'vs/base/browser/dom';
+import { asCSSUrl, ModifierKeyEmitter } from 'vs/base/browser/dom';
 import { domEvent } from 'vs/base/browser/event';
 import { IAction, Separator } from 'vs/base/common/actions';
-import { IdGenerator } from 'vs/base/common/idGenerator'; // {{SQL CARBON EDIT}}
-import { IDisposable, toDisposable, MutableDisposable, DisposableStore, dispose } from 'vs/base/common/lifecycle'; // {{SQL CARBON EDIT}}
+import { IDisposable, toDisposable, MutableDisposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { localize } from 'vs/nls';
 import { ICommandAction, IMenu, IMenuActionOptions, MenuItemAction, SubmenuItemAction, Icon } from 'vs/platform/actions/common/actions';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
@@ -79,10 +78,6 @@ export function fillInActions(groups: ReadonlyArray<[string, ReadonlyArray<MenuI
 		secondaryBucket.unshift(...overflow, new Separator());
 	}
 }
-
-const ids = new IdGenerator('menu-item-action-item-icon-'); // {{SQL CARBON EDIT}} - add back since custom toolbar menu is using below
-
-const ICON_PATH_TO_CSS_RULES = new Map<string /* path*/, string /* CSS rule */>(); // {{SQL CARBON EDIT}} - add back since custom toolbar menu is using below
 
 export class MenuEntryActionViewItem extends ActionViewItem {
 
@@ -262,150 +257,3 @@ export function createActionViewItem(instaService: IInstantiationService, action
 		return undefined;
 	}
 }
-
-// {{SQL CARBON EDIT}} - This is here to use the 'ids' generator above
-// Always show label for action items, instead of whether they don't have
-// an icon/CSS class. Useful for some toolbar scenarios in particular with
-// contributed actions from other extensions
-export class LabeledMenuItemActionItem extends MenuEntryActionViewItem {
-	private _labeledItemClassDispose?: IDisposable;
-
-	constructor(
-		public _action: MenuItemAction,
-		@IKeybindingService labeledkeybindingService: IKeybindingService,
-		@INotificationService protected _notificationService: INotificationService,
-		private readonly _defaultCSSClassToAdd: string = ''
-	) {
-		super(_action, labeledkeybindingService, _notificationService);
-	}
-
-	updateLabel(): void {
-		if (this.label) {
-			this.label.innerText = this._commandAction.label;
-		}
-	}
-
-	// Overwrite item class to ensure that we can pass in a CSS class that other items use
-	// Leverages the _defaultCSSClassToAdd property that's passed into the constructor
-	protected _updateItemClass(item: ICommandAction): void {
-		dispose(this._labeledItemClassDispose);
-		this._labeledItemClassDispose = undefined;
-
-		if (ThemeIcon.isThemeIcon(item.icon)) {
-			// TODO
-		} else if (item.icon) {
-			let iconClass: string;
-
-
-			if (item.icon?.dark?.scheme) {
-				const iconPathMapKey = item.icon.dark.toString();
-
-				if (ICON_PATH_TO_CSS_RULES.has(iconPathMapKey)) {
-					iconClass = ICON_PATH_TO_CSS_RULES.get(iconPathMapKey)!;
-				} else {
-					iconClass = ids.nextId();
-					createCSSRule(`.codicon.${iconClass}`, `background-image: ${asCSSUrl(item.icon.light || item.icon.dark)} !important`);
-					createCSSRule(`.vs-dark .codicon.${iconClass}, .hc-black .codicon.${iconClass}`, `background-image: ${asCSSUrl(item.icon.dark)} !important`);
-					ICON_PATH_TO_CSS_RULES.set(iconPathMapKey, iconClass);
-				}
-
-				if (this.label) {
-					const iconClasses = iconClass.split(' ');
-					if (this._defaultCSSClassToAdd) {
-						iconClasses.push(this._defaultCSSClassToAdd);
-					}
-					this.label.classList.add('codicon', ...iconClasses);
-					this._labeledItemClassDispose = toDisposable(() => {
-						if (this.label) {
-							this.label.classList.remove('codicon', ...iconClasses);
-						}
-					});
-				}
-			}
-		}
-	}
-
-	dispose(): void {
-		if (this._labeledItemClassDispose) {
-			dispose(this._labeledItemClassDispose);
-			this._labeledItemClassDispose = undefined;
-		}
-
-		super.dispose();
-	}
-}
-
-/**
- * This is a duplicate of LabeledMenuItemActionItem with the following exceptions:
- * - Adds CSS class: `masked-icon` to contributed actions label element.
- * - Adds style rule for masked-icon.
- */
-export class MaskedLabeledMenuItemActionItem extends MenuEntryActionViewItem {
-	private _labeledItemClassDispose?: IDisposable;
-
-	constructor(
-		public _action: MenuItemAction,
-		@IKeybindingService labeledkeybindingService: IKeybindingService,
-		@INotificationService protected _notificationService: INotificationService,
-		private readonly _defaultCSSClassToAdd: string = ''
-	) {
-		super(_action, labeledkeybindingService, _notificationService);
-	}
-
-	updateLabel(): void {
-		if (this.label) {
-			this.label.innerText = this._commandAction.label;
-		}
-	}
-
-	// Overwrite item class to ensure that we can pass in a CSS class that other items use
-	// Leverages the _defaultCSSClassToAdd property that's passed into the constructor
-	protected _updateItemClass(item: ICommandAction): void {
-		dispose(this._labeledItemClassDispose);
-		this._labeledItemClassDispose = undefined;
-
-		if (ThemeIcon.isThemeIcon(item.icon)) {
-			// TODO
-		} else if (item.icon) {
-			let iconClass: string;
-
-
-			if (item.icon?.dark?.scheme) {
-				const iconPathMapKey = item.icon.dark.toString();
-
-				if (ICON_PATH_TO_CSS_RULES.has(iconPathMapKey)) {
-					iconClass = ICON_PATH_TO_CSS_RULES.get(iconPathMapKey)!;
-				} else {
-					iconClass = ids.nextId();
-					createCSSRule(`.codicon.masked-icon.${iconClass}::before`, `-webkit-mask-image: ${asCSSUrl(item.icon.light || item.icon.dark)}`);
-					createCSSRule(`.codicon.masked-icon.${iconClass}::before`, `mask-image: ${asCSSUrl(item.icon.light || item.icon.dark)}`);
-					ICON_PATH_TO_CSS_RULES.set(iconPathMapKey, iconClass);
-				}
-
-				if (this.label) {
-					const iconClasses = iconClass.split(' ');
-					if (this._defaultCSSClassToAdd) {
-						iconClasses.push(this._defaultCSSClassToAdd);
-					}
-					this.label.classList.add('codicon', ...iconClasses);
-					this.label.classList.add('masked-icon', ...iconClasses);
-					this._labeledItemClassDispose = toDisposable(() => {
-						if (this.label) {
-							this.label.classList.remove('codicon', ...iconClasses);
-						}
-					});
-				}
-			}
-		}
-	}
-
-	dispose(): void {
-		if (this._labeledItemClassDispose) {
-			dispose(this._labeledItemClassDispose);
-			this._labeledItemClassDispose = undefined;
-		}
-
-		super.dispose();
-	}
-}
-// {{SQL CARBON EDIT}} - End
